+++ core/src/main/scala/com/funlabyrinthe/core/graphics/Image.scala
@@ -3,8 +3,8 @@ package com.funlabyrinthe.core.graphics
 trait Image {
   def isComplete: Boolean
 
-  def width: Double
-  def height: Double
+  def width: Int
+  def height: Int
   def isAnimated: Boolean
 
   /** For a fixed image, 0; for a frame, its delay; for an animated image, its total time. */
