+++ core/src/test/scala/com/funlabyrinthe/core/pickling/FakeGraphicsSystem.scala
@@ -3,10 +3,10 @@ package com.funlabyrinthe.core.pickling
 import com.funlabyrinthe.core.graphics.*
 
 class FakeGraphicsSystem extends GraphicsSystem:
-  def createCanvas(width: Double, height: Double): Canvas =
+  def createCanvas(width: Int, height: Int): Canvas =
     throw UnsupportedOperationException("FakeGraphicsSystem.createCanvas")
 
-  def createFrameCanvas(width: Double, height: Double, time: Int): Canvas =
+  def createFrameCanvas(width: Int, height: Int, time: Int): Canvas =
     throw UnsupportedOperationException("FakeGraphicsSystem.createFrameCanvas")
 
   def createAnimated(frames: List[Canvas]): Image =
